refactor(IgxSelect): Move public exports to index barrel #4182
