+++ projects/igniteui-angular/src/lib/select/index.ts
@@ -1,0 +1,5 @@
+export * from './select-group.component';
+export * from './select-item.component';
+export * from './select.component';
+export * from './select.module';
+
