+++ projects/igniteui-angular/src/lib/select/select.module.ts
@@ -1,0 +1,26 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { FormsModule, ReactiveFormsModule } from '@angular/forms';
+import { IgxIconModule } from '../icon/index';
+import { IgxDropDownModule } from './../drop-down/index';
+import { IgxToggleModule } from './../directives/toggle/toggle.directive';
+import { IgxRippleModule } from '../directives/ripple/ripple.directive';
+import { IgxInputGroupModule } from '../input-group/input-group.component';
+import { IgxButtonModule } from '../directives/button/button.directive';
+
+import { IgxSelectComponent, IgxSelectToggleIconDirective } from './select.component';
+import { IgxSelectItemComponent } from './select-item.component';
+import { IgxSelectItemNavigationDirective } from './select-navigation.directive';
+import { IgxSelectGroupComponent } from './select-group.component';
+
+/** @hidden */
+@NgModule({
+    declarations: [IgxSelectComponent, IgxSelectItemComponent, IgxSelectItemNavigationDirective,
+        IgxSelectToggleIconDirective, IgxSelectGroupComponent],
+    exports: [IgxSelectComponent, IgxSelectItemComponent, IgxSelectItemNavigationDirective,
+         IgxSelectToggleIconDirective, IgxSelectGroupComponent],
+    imports: [IgxRippleModule, CommonModule, IgxInputGroupModule, FormsModule, ReactiveFormsModule,
+        IgxToggleModule, IgxDropDownModule, IgxButtonModule, IgxIconModule],
+    providers: []
+})
+export class IgxSelectModule { }
