+++ projects/igniteui-angular/src/public_api.ts
@@ -80,9 +80,7 @@ export * from './lib/tabbar/tabbar.component';
 export * from './lib/tabs/index';
 export * from './lib/time-picker/time-picker.component';
 export * from './lib/toast/toast.component';
-export * from './lib/select/select.component';
-export * from './lib/select/select-item.component';
-export * from './lib/select/select-group.component';
+export * from './lib/select/index';
 
 
 /**
